compensation/requirements.txt
asgiref==3.7.2
black==23.11.0
certifi==2023.11.17
chardet==3.0.4
click==8.1.7
colorama==0.4.6
Django==3.2.17
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.14.0
elastic-transport==8.10.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.0
idna==2.10
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.24.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.25.11